refactor: moving lineage code into its own plugin
There are a number of open source data governance platforms available.
Some of the most popular ones are:
* Amundsen: https://www.amundsen.io/
* Atlas: https://atlas.apache.org/
* Marquez: https://marquezproject.github.io/marquez/

They all have data lineage functionality that is exposed via APIs.

This refactor will enable user to replace the default lineage backend.
