+++ querybook/config/querybook_default_config.yaml
@@ -13,6 +13,9 @@ REDIS_URL: ~
 ELASTICSEARCH_HOST: ~
 ELASTICSEARCH_CONNECTION_TYPE: naive
 
+# --------------- Lineage ---------------
+DATA_LINEAGE_BACKEND: logic.lineage
+
 # --------------- Database ---------------
 DATABASE_CONN: ~
 DATABASE_POOL_SIZE: 10
