+++ querybook/server/datasources/metastore.py
@@ -18,6 +18,7 @@ from lib.metastore.utils import DataTableFinder
 from lib.query_analysis.samples import make_samples_query
 from lib.utils import mysql_cache
 from logic import metastore as logic
+from logic.metastore import lineage
 from logic import admin as admin_logic
 from models.metastore import (
     DataTableWarning,
@@ -118,7 +119,7 @@ def add_data_job_metadata(
         )
         table_lineage_ids = None
         if query_text:
-            table_lineage_ids = logic.create_table_lineage_from_metadata(
+            table_lineage_ids = lineage.create_table_lineage_from_metadata(
                 data_job_metadata.id
             )
         return {
@@ -307,7 +308,7 @@ def get_table_query_examples(
             environment_id, session=session
         )
         engine_ids = [engine.id for engine in engines]
-        query_logs = logic.get_table_query_examples(
+        query_logs = lineage.get_table_query_examples(
             table_id,
             engine_ids,
             uid=uid,
@@ -328,7 +329,7 @@ def get_table_query_examples_users(table_id, environment_id, limit=5):
     verify_data_table_permission(table_id)
     engines = admin_logic.get_query_engines_by_environment(environment_id)
     engine_ids = [engine.id for engine in engines]
-    users = logic.get_query_example_users(table_id, engine_ids, limit=limit)
+    users = lineage.get_query_example_users(table_id, engine_ids, limit=limit)
 
     return [{"uid": r[0], "count": r[1]} for r in users]
 
@@ -337,7 +338,7 @@ def get_table_query_examples_users(table_id, environment_id, limit=5):
 def get_table_query_examples_concurrences(table_id, limit=5):
     api_assert(limit <= 10)
     verify_data_table_permission(table_id)
-    concurrences = logic.get_query_example_concurrences(table_id, limit=limit)
+    concurrences = lineage.get_query_example_concurrences(table_id, limit=limit)
     return [{"table_id": r[0], "count": r[1]} for r in concurrences]
 
 
@@ -513,7 +514,7 @@ def add_lineage(table_id, parent_table_id, job_metadata_id):
         Returns new table lineage dictionary
     """
     with DBSession() as session:
-        table_lineage = logic.add_table_lineage(
+        table_lineage = lineage.add_table_lineage(
             table_id, parent_table_id, job_metadata_id, True, session=session
         )
         return table_lineage.to_dict()
@@ -522,13 +523,13 @@ def add_lineage(table_id, parent_table_id, job_metadata_id):
 @register("/lineage/<int:table_id>/parent/", methods=["GET"])
 def get_table_parent_lineages(table_id):
     with DBSession() as session:
-        return logic.get_table_parent_lineages(table_id, session=session)
+        return lineage.get_table_parent_lineages(table_id, session=session)
 
 
 @register("/lineage/<int:table_id>/child/", methods=["GET"])
 def get_table_child_lineages(table_id):
     with DBSession() as session:
-        return logic.get_table_child_lineages(table_id, session=session)
+        return lineage.get_table_child_lineages(table_id, session=session)
 
 
 @register("/table_warning/", methods=["POST"])
